move simple value tests into dedicated file
